Removing endpoint's operation-specific methods
